pkg/gitserver: Supply git credentials via credential helper, add ability to specify username. (#1869)
Supply credentials via credential helper rather than GIT_ASKPASS.

Credential helper is a more flexible, simple and reliable mechanism to
provide credentials to git. Critically, it enables providing username
as well as passwords in the constrained environment of running a git
command.

The behavior of credential helpers and GIT_ASKPASS is documented at
https://www.kernel.org/pub/software/scm/git/docs/gitcredentials.html.

Add ability to provide username to gitserver. This is going to be needed
for GCP integration, since their private repositories require both a
username and password to git clone.

Remove support for Windows. It's not needed at this time. Can bring
back when there's a need.

Document purpose of separate file with credentials data and
executable helper.