client/browser-ext: always wait at least HOVER_TIME before showing hover and 3*HOVER_TIME before "Loading..." (fixes #1891)
In the Chrome ext, the "Loading..." hover shows up nearly
immediately, which results in a jittery feel if the actual hover shows
up shortly thereafter.

For example, this timeline feels jittery because there is a quick
flash of the initial "Loading..." text:

* t=0 Loading...
* t=60msec var tags []string

To avoid this on the main site, we only show Loading... if the hover
data has not loaded within a certain time period (600msec, or 3 times
the hover delay). If the hover data loads within 600msec, no
"Loading..." is shown; only the real hover text is shown.

This change adds this same behavior to the Chrome ext.