Add ability to treat repo already exists state as 200 status. (#1898)
Certain clients, such as browsers, don't have the ability to make
requests that result in 4xx status codes, without also displaying a
scary red log entry in the console. Provide a way for them to call the
create repo method without first checking if the repo exists. The
result becomes 200 OK rather than 409 Conflict, and the result is
empty. This is fine for the current needs of such a client (the chrome
extension).

The goal of ensureRepoExists function is to do as its name says. If a
repo already exists, that's an acceptable success state. Use
AcceptAlreadyExists=true option to prevent 409 Conflict status code
from being returned, which we have no way to hide here (unlike 200 OK,
which is not displayed).

Fixes issue where users may be mislead by seeing a red 409 error
message on a POST request to /.api/repos endpoint, and not realize this
is an intended/acceptable result.