No need to generate Google API package in CI.
It has a //go:generate directive, but that's meant for local generation.

It can be generated each time, but there's simply no point and no
value. It can cause issues if google-api-go-generator changes. So, best
to stick with manual occasional generation whenever needed.

Eventually this package will be available from official Google API
package, and then we won't need our own copy of it here.