ui: Add UI for Google auth (behind feature flag). (#1905)
This change includes support for future Google Cloud Platform
repositories, which requires the ability to initiate linking a Google
account. This is done via a button similar to adding private GitHub
repositories.

This button is not displayed, unless opted in via a feature flag
by setting localStorage["google"] = true in frontend.

This commit builds on work done in #1930.