ui: Normalise pos used for caching on hover/j2d (#1952)
* ui: Normalise position based on word before sending hover/j2d

This will result in more consistent behabiour w.r.t. to how we cache. At the
moment, we cache the very first position we discover on a word, which can lead
to some perculiarities due to positions that the editor model can return.
Instead we now always send requests as if hovering over the middle of a word.

Also changed was that the cache key is never null, since we can just fallback to
using the input position if there is no word at a position.

* ui: Remove always true conditionals

Due to the previous change, key is always non-null. So we can simplify the
conditionals for key.