pkg/gitserver: Always provide username to git, even if blank.
If the username is not provided to git via credential helper, it will
ask via stdin or fall back to other methods. It's better to directly
provide it for all values.

Previously, having blank username would cause the clone operation to
block waiting on stdin. After this change, git uses the blank username
and sends it to remote server.