dev: do not auto-install dev langserver (#2002)
Previously, `make serve-dev` auto-installed the JS/TS lang server. It
did not make it clear how to update it, and it would fail `make
serve-dev` if the JS/TS lang server repo clone was in a bad state
(e.g., `yarn` was killed halfway through its work).

This makes `make serve-dev` no longer auto-install any dev lang
servers (except the builtin Go lang server). Running `make serve-dev`
now prints out a message explaining how to install the JS/TS lang
server if you so choose. If you have already installed the JS/TS dev
lang server, it tells you how to update it (to make it clear that it
does not auto-update).

It is important to make it easy to install the dev JS/TS lang server
(and future dev lang servers), so that our devs don't only try out
stuff in Go. (Most changes should not affect whether Go or TS/JS is
used, but they occasionally have different behavior, such as docstring
formatting on hover and def landing.)

When we have many more lang servers, we'll have to come up with a more
scalable approach for maintaining/installing them on a dev machine,
but this is fine for now. (We could have dev servers just talk to the
LSP servers in prod through the existing API. It would take a bit of
additional work but it's doable.)