xlang: Record and log server proxy stats on idle shutdown (#2040)
* xlang: Move serverProxyConn.last into serverProxyConnStats

I want to start instrumenting more metrics for connections to an LSP server. The
first I want to expose is Last. So this commit introduces the
serverProxyConnStats struct, as well as moves last into it.

This is following a similiar pattern to the Stats call in the `database/sql`
package.

* xlang: Log server stats when done with server

* xlang: Record call counts in server proxy stats

* xlang: Record to prom serverProxyConn shutdown stats