Revert "services/ext/github: Only update rateLimitRemainingGauge when header present."
This reverts commit 3f2709356f6a73d1d96ef407948597c1082f4cae and efde4682fbaa25f078ea49c1781e653eb2a86859.

That change was done in order to collect information from production
and test out a theory about the source of sudden temporary spikes of
0 rate limit remaining values. It was a theory that it was caused by
that header missing, but the data collected with this change shows
that it's not. There must be another source of that.

Revert this change since it's no longer neccessary.