xlang: Retry transient errors from server (#2073)
If our backend dies or we have temporary network issues, we can just retry. This
implementation is relatively conservative, and sleeps at most for 1 second (in
theory we could wait a bit longer).

We retry at this layer since it will potentially start up a new LSP server if we
need to. It should also nicely interact with our tracing instrumentation.

This is our 2nd most common reason to fail user requests. This is most likely
due to our current OOM issues.