pkg/gitserver: Improve clarity, documentation. (#2086)
This is a zero behavior difference refactor, with the goal of improving
readability of this code.

It was a mistake to make setSpanTags a method of request, because that
makes it sound like it performs some operation on the request object.
Instead, it operates on the span, setting tags _from_ the request.

Without looking closely at that, when seeing code like:

	req, _ := newRequest()
	req.setSpanTags(span)

It makes it look like there's a mistake in the code, because req is
actually discarded, so perhaps setSpanTags has no effect. When
rewritten as a function that takes span as a first parameter:

	req, _ := newRequest()
	setSpanTags(span, req)

It becomes more clear that it's fine for req to be discarded, since
it's only needed to set span tags on span.

Also fill in missing documentation for request types.

Updates #1546.