added github_access_token_override
This adds the optional field github_access_token_override to a user's app_metadata. If this field is set, then the GitHub token returned by the OAuth flow is ignored and this token is added to the session instead. This will be used for implementing read-only access to private repos via dummy user.