services/backend/internal/localstore: improve legacy total refs counting
Picking an arbitrary def_key to perform the total repo ref selection on is not good
because the first ref might not be the most referenced thing. By selecting the best
of the first 10,000 we often get a better sample (e.g. 909 -> 36k for golang/go).
Perf is pretty good, too, ~36ms for golang/go.