{xlang/...}: rewrite `dir:` `workspace/symbol` queries for github.com/golang/go
In ALL cases an expanded (i.e. inclusive of `vendor`) Go package path, when aligned
relative to `$GO[PATH,ROOT]/src/<Initialize.RootPath>`, is equal to that of the `dir:`
filter of `workspace/symbol`. However, because `github.com/golang/go` is not a valid
Go package and requires special-casing (i.e. it's sources are in `github.com/golang/go/src/`
yet the actual packages are `$GOROOT/src/`) we must deal with this somewhere.

It is incorrect to place this handling inside the Go language server, as I did in https://github.com/sourcegraph/go-langserver/pull/28
because the only situation in which this can occur is `github.com/golang/go` due to how
we special case it. Therefor, the problem is with the build server not the language
server. This became evident after trying to test for this case in the language server:
it is not possible.

In contrast, when implemented this way, it makes much more sense and is testable clearly
as part of the same codebase (see `integration_test.go`).

See also related PR https://github.com/sourcegraph/go-langserver/pull/30