xlang: Record user errors differently (#2363)
We shouldn't make the distinction of user error, since our
frontend code shouldn't send "invalid" requests. An example is
the browser-ext sending hover requests for private repos we are
not authorized to view. For now we will skip recording user
errors in our apdex scores, but record them in honeycomb for deep
diving.