xlang: PHP (#2419)
* xlang: Valid RootPath for PHP

The PHP langserver follows the spec, so we need to do a few updates to get it to
work correctly. We should update everything to use a path, since we always just
pass in `file:///`

* ui: Enable PHP support in the frontend

* xlang: Document that only PHP follows rootPath spec

* ui: Only enable PHP if phpLangEnabled is set in localStorage

As part of the change I unexported `supportedExtensions` as well as making it a
`Set<string>`.

You can enable with `localStorage.setItem("phpLangEnabled", "true");`

* dev: Support PHP from make serve-dev

* dev: Cleanup log messages for local language servers

We don't need to repeat ourselves about using a local langserver (once from
bash, another time from xlang). Also shortern verbose message about xlang
managing an executable.

* ui: clean up php feature flag code