golang: Just in time dependency fetching (#2761)
* govendor update github.com/sourcegraph/go-langserver/langserver

* xlang: bctx does not need a read guard

https://github.com/sourcegraph/go-langserver/pull/59 introduced a better
underlying FS which is safe for concurrent access. As such we can just directly
use the FS, rather than doing a copy on read each time we want to do a
bctx.Import.

* golang: Single flight dependency fetching across pkgs

Currently we would duplicate work across pkgs, and potentially mount a
dependency more than once onto the FS. This moves the dep url locking to the
BuildHandler from being local to the dependency analysis for a package.

* golang: Extract dependency aware importPackage into findPackage

We want to use this function in other places.

* golang: Use the build servers findPackage to typecheck

This enables just in time fetching of dependencies. As such we can avoid the
separate step of fetching them before typechecking.

* golang: reset all buildserver fields

This looks like a latent bug. However, we don't actually ever re-use a build
server, so the bug is benign in practice.

* golang: Fetch github.com/golang/go on startup

This hides the latency of downloading the golang repo on the first ever
typecheck a buildserver process does.

* final govendor update to match go-langserver master