Support resolving vanity imports like gopkg.in (#2790)
* golang: Remove resolvedPath field

It is unused, and based on the test data it was returning inconsistent values.
Rather just remove it.

* golang: Add better support for gopkg.in

This will fetch gopkg.in via github, rather than using gopkg.in's git proxy.
This has the advantage of being faster, and in the future allowing us to better
support j2d for gopkg.in dependencies.

* golang: Use mounted dependencies for url rewriting

Previously we would have some hardcoded strings to convert path to dependencies
into urls which allow us to jump across repos. However, we control everything
that appears on the VFS, so we simply just need to keep track of the mounted
dependencies and then consult the list to rewrite the location of a file. This
commit implements that.