golang: Integration test for glide support (#2806)
* golang: Integration test for glide support

* xlang: Ensure integration test cleans up

Previously we did not close the listener, nor did we close the pipes. This left
goroutines running which referenced our proxy, etc. This is likely a big reason
why the xlang integration test used so much memory, since we had background
goroutines still holding on to references.

Now when you run the integration test, at the end of the test there are no xlang
goroutines running.