xlang/golang/buildserver: optimize workspace/xreferences BSP method (use hints)
This causes the workspace/xreferences method to only fetch transitive dependencies
for the directory that references were found in (by workspace/xdependencies, stored
in the DB, and subsequently passed into xreferences).

This makes dependency fetching for workspace/xreferences go from >60s for github.com/kubernetes/kubernetes
down to a more reasonable 3s to <1s or so (I did not measure exactly).