services/backend/internal/localstore: allow Dependencies method to return multiple per-repository dependencies
We used to have `distinct on (repo_id)` from when we only had one record per
repository per dependency in the DB. Since we've listed this restriction (i.e.
we can get back multiple entries with different `hints` results).

We didn't notice this before because workspace/xreferences didn't respect hints,
and thus as long as we got a record for a repository -- all dependencies would
be searched. This is no longer the case, so enable returning multiple results
per repository.