xlang/golang/buildserver: add a request-scoped cache to dynamic import path resolution
This improves the performance of transitive dependency fetching for `workspace/xreferences`
and `workspace/xdependencies` only. Due to they way they must be structured, `fetchMeta` is
invoked multiple times for the same import path. This becomes quite a bottleneck as proven
with the results below.

This was benchmarked with many (but not all) directories of each repository
included in the `dirs` hint (a real `peek external refs` query):

| Repository                         | Old   | New   | Delta   |
|------------------------------------|-------|-------|---------|
| `github.com/gorilla/csrf`          | 2.4s  | 2.3s  | -4.16%  |
| `github.com/gorilla/websocket`     | 2.7s  | 2.5s  | -7.40%  |
| `github.com/kubernetes/kubernetes` | 35.4s | 17.1s | -51.69% |