golang: normalise srcDir and cache FindPackage (#2832)
Currently most of the work for typechecking is repeated work due to having to
take into account vendoring. With this change, we avoid the repeated work by
doing the fast vendor checks which FindPackage will do anyways, but then caching
based on those checks.

Locally this takes the kubernetes integration test from on average `6.2s` to `3.9s`.

This makes a big assumption, that the FS is the same as the build servers FS.
This is currently a correct assumption. There is potential for race conditions
when we start using the overlay, but that can be addressed later.

This is also the sort of change the could be put upstream into the `loader` pkg.
However, doing it this way we get to share FindPackage results across packages.
Next I am keen to experiment with removing the typecheck cache and just rely on
the FindPackage cache. I have a feeling that will greatly help with our memory
usage.