rcache: Do not namespace by hostname/app url (#3074)
We now run a redis per namespace, so we do not need to do this. Our main usecase
of rcache is http caching. Before this change different microservices didn't
share the http cache due to SRC_APP_URL only being set for sourcegraph-frontend.

Note: When this is deployed it will invalidate the cache. It should be deployed
when traffic is not high.