services/httpapi: add mode to "LSP HTTP gateway" span names + RootPath tag
Before we had span names like this:

```
LSP HTTP gateway: workspace/xreferences
```

This is problematic because HTTP gateway could be for any language, so
creating a Lightstep dasboard for this encompasses all languages instead
of just measuring one language. Change it to include the language:

```
LSP HTTP gateway: go: workspace/xreferences
```

So that each language can be tracked independently at the level closest
to the actual client request.

Additionally, tag each span with `RootPath` so it becomes easier to see
at a high level which repositories tend to be in e.x. our p99 category.