allow querying global refs of definition called in private repos (#3315)
This relaxes the access control check to allow you to see global refs for (e.g.) a call to strings.SplitN (a public def) in a private repo.

Prior to this change, if you were viewing a private repo, you could never see global refs.

It is still the case that global refs will not return calls from a private repo to any definition. This change does not affect that.