ui: stream global refs into WOW sidebar
This code allows us to effectively poll on an interval for global refs results
to populate in the sidebar, without spamming xreferences. It is configured to
poll until non-empty results have been fetched for 5 dependent repos. The polling
is disposed as soon as the WOW sidebar is dismissed so we don't put unnecessary load
on the server.

/cc #3325