xlang: use the initializationOptions custom field for "mode", not a top-level field
The LSP spec defines an any-typed initializationOptions field in the initialize params: https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#initialize-request. We should use that instead of our own top-level field.

This change is backward compatible, so it won't break if frontend and xlang are on different versions. The HTTP API handler for xlang sets both fields in the initialize params it sends to the xlang client proxy server, and the xlang client proxy server accepts the field value in either field.

(This change will let us reuse more TypeScript code for LSP support in the future, and it's good practice in general.)