ui: consolidate WOW feature flags
One side effect is that code lens state is not persisted to local storage. I'd argue this is actually a feature, not a bug,
because having authorship enabled when loading a blob URL file (at least currently) causes slowness/jitter.

If we do want to persist code lens state across sessions, we need to do that somewhere else besides the Features module.