{services/backend,xlang}: do not hit DB or xreferences for non-referenceable symbols
Prior to this change we would query the DB and roundtrip many times for xreferences
(which is very slow for Go, because the xreferences implementation is not smart in
this regard). This change enables us to bypass querying the DB and invoking xreferences
entirely by asking languages to define as part of SymbolDescriptor whether or not
the symbol can be referenced.

This is only an optimization, so languages technically do not have to implement it
if it is troublesome to do so for any reason.