xlang: save diagnostics for clients that connect after the server originally published them (#3363)
* xlang: save diagnostics for clients that connect after the server originally published them

Consider the following scenario:

1. Alice opens a Go file with diagnostics on Sourcegraph and receives the diagnostics (which she sees as red squigglies)
2. Bob opens the same Go file 1 minute later

Bob will not see the diagnostics (and no red squigglies) because the Go lang server sent the textDocument/publishDiagnostics 1 minute ago, and it only sends it when it performs an uncached typecheck operation (and other uncached operations). If the operations used to satisfy Bob's requests are cached, he won't receive the diagnostics because the lang server doesn't (and has no way to) resend them.

This change fixes that by storing the last diagnostics for each file for each lang server. Clients that connect after the diagnostics are sent receive the diagnostics when they send a textDocument/didOpen notification for the file.

This has zero impact on our current UX and API traffic because we do not ever send textDocument/didOpen. If a client does send textDocument/didOpen, it will receive prior notifications, but there are no bad security implications because it could have received those if it happened to be the first client to initialize that lang server.

* xlang: use feature flag to let us safely measure perf impact of diagnostics saving

* address CR comments