#3461 – Remove transition on filetree (#3620)
* Remove transition altogether.

As far as I can tell, there's not a good way to remove the transition
from :hover, but keep it on the selected class change. Just remove.

Sorry about this, 500ms was never supposed to go it. It was supposed to
be 100ms, but since it doesn't work on selected, let's just remove.

* Remove animation from other trees.

* Add back Tree.css (only remove tree.css)

Weird my local tests didn't catch this