xlang: Stream results from langserver back to httpapi (#3687)
* govendor fetch github.com/sourcegraph/go-langserver/...

* golang: Proxy partialResult notify

This allows the buildserver to the notification from the language server onto
lsp-proxy.

* golang: Proxy ID to langserver

The language server needs to know the ID so `$/partialResult` requests contain a
valid ID. We could avoid passing the ID, and instead insert it in the
jsonrpc2ConnImpl. However, not passing the ID is hiding a bug that doesn't
affect us yet. `$/cancelRequest` depends on the ID, but we don't use it yet
(will soon though).

* xlang: Include a client request id in server request IDs

This allows us to tie back requests for a client to a language server.
We need this to stream partial results back to the correct client. To make this
possible we had to introduce a client ID (unique to a process lifetime).

* xlang: Proxy and rewrite partialResult notifications

This proxies notifications we receive from servers back to clients. We also to
take care to rewrite URLs (located in the patches to lsp.Location) and the ID.
We now pass the original client request ID all the way down to the language
server, so it is contained in the partial result params ID.