xlang: Use tmpfriend for lsp-proxy and xlang-go
tmpfriend removes tmp folders from stale processes. I have a suspicion that when
xlang-go or lsp-proxy get into a loop of failed startups, it has something to do
with disk space. However, even if I am wrong, this is just good practice to do
since we write to /tmp but never clean up.