services/backend/internal/localstore: pack database handle into ctx
At some point `dbstore_db_test.go` `testContext` function was modified
to, instead of storing the database handle in the context, just directly
modify the global app database handle. This is fine, except in cases
where two tests are running at the same time. Generally, this isn't
noticed because "one" of those writes wins and both tests use the same
database -- but this isn't always the case as we saw with `pkgs_test.go`
tests as we saw with https://github.com/sourcegraph/sourcegraph/issues/3448#issuecomment-274713729