{pkg/testdb,services}: do not attempt reuse of 'pristine' databases
Doing so leads to strange race issues such as queries that are in flight
returning old (non-truncated) database data while truncation operations
are in flight. Besides, database creation is fast assuming we do it ahead
of time / don't block the test on it. This simplifies quite a lot of code.