services/backend/internal/localstore: manually map repo -> import path
This fixes issue https://github.com/sourcegraph/sourcegraph/issues/3924 by
creating a manual (hand written) map of popular Go custom import paths into
their repository prefix counterparts, so that the TotalRefs endpoint works
on them.

It is needed for the badges release tomorrow, and is implemented as a workaround
for custom import paths in lieu of having a proper mapping (which will be the
`pkgs` table in the future, but it does not include Go data right now).