httpapi: Proxy request ID (#4007)
Before this change we would assign our own IDs to talk to the backend, but
rewrite the response to match the ID. However, if we receive a request like
`$/cancelRequest` it specifies an ID. To support that we would have to rewrite
the ID to match the ID we used. Instead, we can just use the same ID, which is
what this change implements.