httpapi: Propagate opentracing in lsp proxy (#4044)
* httpapi: Add timeout to dialing lsp proxy

We use the same timeout that the HTTP gateway used.

* httpapi: Propagate opentracing in lsp proxy

xlang.Client will do that for us. This has the downside that every request part
of the websocket session will have the same root. I will solve this in a future
commit.

* httpapi: Counter for infra errors in lsp

Also removed some noisey which would happen due to user error.

* add in missing Inc, remove logspam