ui/web_modules/sourcegraph/util: properly display external references when there is only one
It's not clear to me why we had checked if there is exactly one reference and consider that
an error. This MUST be a bug. Probably someone typo'd `=` instead of `<` -- or they thought
that there would always be one reference (the repo itself where the definition is defined),
but that isn't always true.

This was an issue uncovered by Sprague, when he was trying to create a test repository for GDS,
and King helped me extensively in tracking this down / pointing me in the right direction since
it was quite confusing.