polyfill es6 iterator usage in vscode-languageclient (#4308)
vscode-languageclient uses some es6 iterator for-of loops that tsc
with target=es5 translates (incorrectly) into for-loops using
`x.length`. tsc considers polyfills for included JS code out of scope,
so as long as we depend on JS libs in node_modules that emit to es6,
we will need to add polyfills to them. See
https://github.com/Microsoft/TypeScript/issues/4031 and
https://github.com/Microsoft/TypeScript/issues/3164.

This is a necessary but not sufficient fix for
https://github.com/sourcegraph/sourcegraph/issues/3906. The `for
(const selector of selectors) {` line in vscode-languageclient was
being unexpectedly (and "incorrectly") compiled from es6 to es5 by our tsc.