xlang: pass along JSON-RPC 2.0 notifs as notifs, do not add ID (#4312)
Our Go jsonrpc2 build/lang servers didn't complain if it received notifications (textDocument/didOpen, etc.) as requests (i.e., if they had an ID field set). But other jsonrpc2 server impls do care; vscode-languageserver cares.

This change makes it so that the LSP proxy does not add an ID to notifications it is proxying. This is necessary for text document change operations to work with our JS/TS langserver.