gobuildserver: Allow LS to invalidate caches on TextDocumentSync (#4318)
* gobuildserver: Use Handle instead of HandleFileSystemRequest

We need to do this since `HandleFileSystemRequest` bypasses the LS's cache
invalidation logic. We also have the benefit of simplifying the code, and just
passing on all notifications directly.

* gobuildserver: Mount workspace after overlay

Before this change, the workspace would override changes in the override. This
would mean essentially all file sync operations would be no-ops. After this
change, changes in the overlay are reflected in the workspace.