xlang: Isolate xreference to bg LS servers (#4353)
We have very strong evidence that xreference on large repos is the culprit
behind a lot of our OOMing. Before the underlying OOM is fixed, we can start to
mitigate the problem by isolating it from the usual LS servers. This makes sense
not just due to isolation, but also due to resource usage. The workspace
lifetime of an xreference request is pretty much once-off, unlike the usual LS
which lives as long as the user session.