xlang: Isolate large repos (#4360)
We are currently demoing sourcegraph on the kubernetes repo. However, the k8s
repo is very large and it being on a shared LS leads to OOMing. This change
allows us to isolate the k8s repo to a specific LS and tweak the resources of
that LS.

In the future this sort of hack shouldn't be needed, but is important for our
short term needs.