ui: use GitTextFileService throughout (even with zap disabled) to unify file access between zap and non-zap (#4431)
Previously, if Zap was enabled, the new GitTextFileService was used. If it was disabled, then a readonly document was provided with the string contents of the file in Git.

This change makes it so that the GitTextFileService is always used. The goal is to reduce the number of different code paths. This also fixes a few bugs, such as the sidebar not showing up when Zap is enabled, and saving not working when Zap is enabled.