httpapi: Connect to lsp-proxy before upgrading to websocket (#4605)
* httpapi: Connect to lsp-proxy before upgrading to websocket

This is the cause of the `response.Write on hijacked connection` log spam. We
could instead write a response to the websocket, but LSP does not have a nice
way to do this. Communicating this via the HTTP response is a cleaner way to do
this. We also expect websocket upgrading to be very rare.

* Revert "improve handling of websocket failure"

This reverts commit 1a4115306460f41721aebcfd601377b5438b1d14.

* govendor fetch github.com/NYTimes/gziphandler/...

* httptrace: Use a httpsnoop to correctly intercept http.ResponseWriter

http.ResponseWriter has several optional interfaces it could implement. Our
naive implementation didn't implement any of those, leading to weirdness.

* httpapi: remove getHijacker hack

Now that we use httpsnoop and have upgraded the gzip handler, the wrappers
around ResponseWriter are good. I also can't reproduce writing to the hijacked
connection anymore, I assume since the response writer is correctly wrapped now.

* httpapi: Supress close error logs for lsp

We still do not supress for zap, but that is much less log spam.

* Do not double write ws upgrade error