xlang: Expire a server if it only has done xref requests (#4746)
We keep implementing this behaviour close to the client layer. However, we have
been iterating a lot on the client recently and keep regressing it. The most
recent regression in this behaviour is more than likely the reason why we are
OOMing at a higher rate in the last 2 weeks. This implementation moves the logic
down into the server_proxy layer, which is the layer actually responsible for
expiring servers. We can now remove our other hacks around this.

To implement this a new stat to track was added, how many requests have been
proxied, but only incremented once the LS has returned. This allows us to expire
a language server if the only request it has done is xrefs, but not expire it
before it has finished processing the request.