gobuildserver: Use glide.yaml's root import path (#4776)
glide.yaml specifies the root import path to use. If a glide.yaml is present, we
can use that to quickly and correctly determine the correct root import path.
Glide is popular in uber repos which are often hosted in github.com for open
source projects, but the actual canonical import is go.uber.org.

There are also enterprise customers who have canonical import paths which have a
suffix of `.git`, which we strip out a few lines before. That is also a bug we
need to fix, but may affect other users. So this is the safer fix to do.