zap: standardize ref naming scheme and resolve Zap revs
A Zap rev is like a Git rev: some fuzzy string that resolves to a commit maybe on a ref (branch).

A Zap ref is like a Git ref: a named thing like a branch or tag (usually a branch).

Resolving a Zap rev can yield any of these 3 outcomes, just like resolving a Git rev:

- Unable to resolve to a revision
- Able to resolve to a revision (a particular snapshot of a workspace) but the rev does not specify which branch, so this is an absolute reference
- Able to resolve to a revision on a branch, so the UI can show additional context about the branch

This commit also updates the names used for variables, fields, etc., to use this distinction between a Zap rev and Zap ref.