searcher: Initial version of searcher service (#4943)
Searcher is a simple service to do text search on a repo at a commit. Its main
purpose is to help avoid load on the gitservers (ie scale out `git grep`).
Currently this is a very simple but working service. It is not yet integrated
into the new textSearch graphql endpoint. However, once it has improved enough
it will replace the `git grep` implementation.