app/installzap: timeout connection to Google Cloud Storage after 30s
Prior to this change we didn't specify a connection timeout when attempting
to download binaries from Google Cloud Storage. This is bad because if we
do not, curl will attempt reconnection constantly with a backoff. Or it can
wait up to 5 minutes do to the default C API dns resolution timeout, depending
on the OS.