searcher: Use a consistent hash of replicas backed by k8s (#5128)
When we have more than one searcher replica running the k8s service will just
round-robin between the different services. Ideally we would want to keep
hitting the same replica for a workspace since it will have a cached copy of the
workspace to search (avoid load and perf hit of hitting gitserver). This
switches us from using the svc loadbalancer to a consistent hash to achieve
that.

I investigated using the nginx ingress since that supports session affinity.
However, it is stateful (it returns a cookie to reuse). This makes it hard for
the sourcegraph-frontend replicas to co-ordinate. Consistent hash allows us to
achieve the same functionality without sharing state. Instead we rely on the
state from the k8s API, which is the same method kube-proxy uses (the service
load balancer).