searcher: Evict items from on disk cache based on total cache size (#5142)
* searcher: Factor out getArchive

This makes future changes clearer.

* searcher: Store includes closer

This is so we can keep track of open archives. We need this functionality when
we implement the LRU.

* searcher: Store archives in an LRU list

We intend to delete archives once the size has reached a limit. This adds the
bookkeeping necessary to know which archives are least recently used.

* searcher: Implement cache eviction

We use a local SSD for caching. It is bounded in size so we need to remove items
from the cache to prevent running out of space. This commit introduces cache
eviction with an LRU algorithm. Notes:

* When to evict is based on the total size of `*.zip` on disk.
* What to evict uses the LRU algorithm.
* To evict we need to wait ongoing searches on the file to finish. We can then
  close the zip and file. However, we can remove the file before then. As such
  bookkeeping is added to keep track of ongoing searches.
* Since we are based on the size of files on disk, we need to add files on disk
  which are not yet tracked in-memory. These files can exist due to restarts in
  the process/pod.

* searcher: Add Store.Start to load missing sooner

This also will make the LRU cache play nicely. When loading missing files they
are put at the front of the LRU. So before this change the very first request
would likely have its archive evicted if the cache was already near max size.

* searcher: Add prometheus instrumentation to Store