ui: Trace LSP calls to the console (#5158)
This is adding back a feature we had before switching to websockets. How it is
presented is trimmed down. Enabled via `features.trace.enable()`. To enable this
we update `WebSocketMessage{Reader,Writer}` to allow multiple listeners. There
is a trace feature in `vscode-jsonrpc2`, but currently it isn't exposed in the
language client. This tracer is also more featureful since it times and links
the response to the request.