xlang: Move archive fetching to go buildserver (#5160)
xlang-go is OOMing regularly due to the large number of active workspaces. It
stores quite a lot in memory per workspace due to keeping all files in memory.
This change switches to gobuildserver directly fetching from gitserver if
gitserver client is configured. Given most workspaces are just for go, this
should greatly reduce the memory requirements on lsp-proxy. It should also
have a performance improvement for most workspaces.

Note: lsp-proxy lazily fetches archives. So we do not need to modify lsp-proxy,
we just avoid asking it for files.