xlang/vfsutil: Experiment with storing archives on disk instead of in memory (#5199)
* xlang/vfsutil: Use a more robust file cache for github

It is also re-useable. This will make it easy for us to use it in ArchiveFS and
we can then not put files into memory.

* xlang/vfsutil: Introduce on-disk gitserver archive VFS

We OOM a lot on lsp-proxy still. Instead, lets not store everything in memory
and instead use the local disk. We can then rely on the FS caches to page out a
file that hasn't been accessed for a while. We also use CACHE_DIR, so in
production the on-disk cache can be backed by a local SSD.

The experiment will be measuring performance. If performance is good, I'll add
cache invalidation (ie not fill up the disk). Part of the perf experiment will
be measuring if xlang-go can go back to LSP based remote FS instead of directly
talking to gitserver.