httpapi: Ensure initialize has run first and succeeded in LSP proxy (#5202)
I have a goal to make lsp-proxy not a single point of failure. This refactors the main client to lsp-proxy such that we know which repo a connection is for before we send any requests to lsp-proxy. We will use that information to pick the lsp-proxy replica similarly to how we do it in searcher.

Note: The frontend may re-use a websocket connection for the same repo, but on another commit. This is why we allow more than one initialize requests.

* httpapi: Factor out authorize check in LSP

* httpapi: Only check security from client -> server

The server -> client does not need the security check.

* httpapi: Use sync handler for client -> server

This will gaurentee ordering of the requests we proxy on from the client. We
want this behaviour just when initialize is received is important.

* httpapi: Ensure initialize has run for lsp proxy

We can also simplify our access to initParams, since we access it in our handler
before spinning up goroutines.